+++ cli/azd/CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Other Changes
 
+## 1.21.2 (2025-11-13)
+
+### Features Added
+
+- [[#6169]](https://github.com/Azure/azure-dev/pull/6169) Promote layered provisioning to beta.
+- [[#6154]](https://github.com/Azure/azure-dev/pull/6154) Add `env` property to service configuration for runtime environment variable management in Container Apps.
+
 ## 1.21.1 (2025-11-11)
 
 ### Features Added